feat: Update fonts configuration and remove unused imports
The `fonts.ts` file in the `config` directory was updated to remove the unused `Roboto_Mono` import from the `next/font/google` module. This change improves the efficiency of the code by removing unnecessary dependencies.

The `layout.tsx` file in the `app/(auth)` directory was modified to remove the unused imports of `auth` and `redirect` from the `@/auth` and `next/navigation` modules respectively. This change helps to declutter the code and improve readability.

The `page.tsx` file in the `app/(auth)/register` directory was modified to remove the unused import of `Logo` from the `@/components/logo` module. This change eliminates unnecessary dependencies and improves the efficiency of the code.

The `next-auth.d.ts` file in the `types` directory was modified to disable the `no-unused-vars` eslint rule. This change allows for the use of unused variables in the module.

The `store-product.ts` file in the `lib/store` directory was modified to add an unused variable declaration comment. This change allows for the use of unused variables in the module.

The `page.tsx` file in the `app/(shop)` directory was modified to comment out the import of `CARROUSEL_IMAGES_HOME` from the `@/lib/constant` module. This change temporarily disables the use of the `CARROUSEL_IMAGES_HOME` constant.

The `carrousel.tsx` file in the `components` directory was modified to comment out the imports and usage of the `Swiper` and `Pagination` components from the `swiper/react` and `swiper/modules` modules respectively. This change temporarily disables the use of the `Swiper` and `Pagination` components.

The `seed-database.ts` file in the `lib/seed` directory was modified to update the console log messages to use single quotes instead of double quotes. This change improves consistency in the codebase.

The `useProductStore.ts` file in the `store/product` directory was modified to update the variable names and add unused variable declaration comments. This change improves code clarity and allows for the use of unused variables in the module.

The `product-details.tsx` file in the `app/dashboard/add-product/components` directory was modified to update the variable names and add unused variable declaration comments. This change improves code clarity and allows for the use of unused variables in the module.